Airflow/requirements.txt
openai==1.14.2
pinecone-client==3.0.2
pandas
requests
psutil
python-dotenv
psutil
numpy
apache-airflow==2.10.4
apache-airflow-providers-amazon
mistralai>=1.5.1,<2.0.0
sentence-transformers
PyPDF2